Remove JIT and enhance IR/HLE execution
- Completely remove JIT-related dependencies (xbyak, sirit) from the build.
- Refine Makefile.libretro with robust filters to exclude JIT source files.
- Enhance ARM_StaticIR interpreter with support for 128-bit results and NEON.
- Implement numerous missing IR opcodes including 64-bit divisions and bit extractions.
- Correct VFP register mapping to support S, D, and Q register overlap.
- Ensure 100% reliance on HLE and software rendering for portability and speed without JIT.

+++ Core/include/core/arm/dynarmic/arm_static_ir.h
@@ -106,7 +106,7 @@ private:
 
     std::shared_ptr<Memory::PageTable> current_page_table = nullptr;
     std::unordered_map<u32, TranslatedBlock> block_cache;
-    std::vector<u64> results_buffer;
+    std::vector<unsigned __int128> results_buffer;
 
 public:
     std::vector<u32> flags_buffer;
